% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aimatApp.R
\name{aimatApp}
\alias{aimatApp}
\title{Run an aimat Shiny app}
\usage{
aimatApp(appname)
}
\arguments{
\item{appname}{A character string specifying the name of the app to run. The app should be located in the `aimat` package's system file directory.}
}
\description{
This function launches a Shiny app for training neural networks using the `aimat` package.
}
\examples{
if (interactive() & requireNamespace("shiny", quietly = TRUE)) {
  aimatApp("app_full")
}
}
