% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bog}
\alias{bog}
\title{Words in an artificial childrens book (in Danish)}
\format{
## `bog`
A vector of words
}
\usage{
bog
}
\description{
Words in an artificial childrens book (in Danish)
}
\keyword{datasets}
