% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fibonacci.R
\name{fibonacci_sphere}
\alias{fibonacci_sphere}
\title{Fibonacci Sphere}
\usage{
fibonacci_sphere(samples = 1000)
}
\arguments{
\item{samples}{Number of points to generate on the sphere.}
}
\value{
A matrix of points on the sphere, where each row is a point in 3D space.
}
\description{
Fibonacci Sphere
}
\examples{
points <- fibonacci_sphere()
library(plotly)
plot_ly(x = points[, 1], y = points[, 2], z = points[, 3], type = "scatter3d", mode = "markers", marker = list(size = 5))
}
