% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_general.R
\name{nn_fun}
\alias{nn_fun}
\title{General neural network with at most two hidden layers}
\usage{
nn_fun(
  formula,
  data,
  weights = NA,
  n_hidden = c(1, 1),
  activation = "Sigmoid",
  eta = 0.01,
  iter = 1000,
  scale = TRUE,
  lossfun = "squared",
  type = "klassifikation",
  trace = FALSE
)
}
\arguments{
\item{formula}{response ~ predictors}

\item{data}{data frame}

\item{weights}{Starting weights}

\item{n_hidden}{integer vector of length 2 with number of neurons in the hidden layers (can be zero)}

\item{activation}{one of "Sigmoid", "Tangenshyperbolsk", "ReLu", "Softsign", "Identitet"}

\item{eta}{learning rate}

\item{iter}{number of iterations}

\item{scale}{logical to scale the input data}

\item{lossfun}{one of "squared", "cross-entropy"}

\item{type}{one of "regression", "klassifikation"}

\item{trace}{logical to save each iteration}
}
\value{
Fitted neural network model as an object of class "nn"
}
\description{
General neural network with at most two hidden layers
}
\examples{
ir <- iris
ir[,1:4] <- scale(ir[,1:4])
fit_ir <- nn_fun(Species ~ ., ir, n_hidden = c(3,5), eta = 0.01, iter = 1000, lossfun = "cross-entropy", activation = "Sigmoid", type = "klassifikation")
}
