% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_general.R
\name{nn_fun_cv}
\alias{nn_fun_cv}
\title{Cross-validation for neural network}
\usage{
nn_fun_cv(..., k = 5)
}
\arguments{
\item{...}{Additional arguments passed to `nn_fun`}

\item{k}{Number of folds for cross-validation}
}
\value{
Mean accuracy across folds
}
\description{
Cross-validation for neural network
}
\examples{
cv <- nn_fun_cv(Species ~ ., iris, n_hidden = c(3,5), eta = 0.01, iter = 1000, lossfun = "cross-entropy")
}
