% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_viz.R
\name{nn_viz}
\alias{nn_viz}
\title{Visualize a neural network}
\usage{
nn_viz(nn)
}
\arguments{
\item{nn}{FIXME}
}
\description{
Visualize a neural network
}
\examples{
ir <- iris
ir[,1:4] <- scale(ir[,1:4])
fit_ir <- nn_fun(Species ~ ., ir, n_hidden = c(3,5), eta = 0.01, iter = 1000,
  lossfun = "cross-entropy", activation = "Sigmoid", type = "klassifikation")
if(requireNamespace("visNetwork", quietly = TRUE)){
  nn_viz(fit_ir)
}

}
