% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwp.R
\name{nwp_make_word_data}
\alias{nwp_make_word_data}
\title{Make next word prediction data}
\usage{
nwp_make_word_data(corpus, context_size, w2v, return = "both")
}
\arguments{
\item{corpus}{Vector of words.}

\item{context_size}{Integer, size of the context window (using this many words to predict the next).}

\item{w2v}{Word2Vec model object. This is a list with three elements: `W1`, `W2` and `vocab`. `W1` and `W2` are matrices of word vectors, and `vocab` is a vector of words.}

\item{return}{Character, one of "both", "words" or "numeric". If "both", returns a list with two elements: `words` and `numeric`. If "words", returns only the `words` element. If "numeric", returns only the `numeric` element. Default is "both".}
}
\value{
A list containing data to train next word prediction. The list contains one or both of the following elements:
- words: a data.frame with `context_size`+1 columns of words where the last column is the target word.
- numeric: a numeric matrix
}
\description{
Make next word prediction data
}
\examples{
bog_w2v <- w2v(bog, win = 1, hidden_dim = 3, epochs = 5, learning_rate = 0.01, verbose = 0)
bog_data <- nwp_make_word_data(corpus = bog, context_size = 2, w2v = bog_w2v)
}
