% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w2v_nn.R
\name{plot.w2v}
\alias{plot.w2v}
\title{Plot word2vec model}
\usage{
\method{plot}{w2v}(x, ..., which = 1)
}
\arguments{
\item{x}{Word2Vec model object (list with W1, W2 and vocab)}

\item{...}{Additional arguments (not used)}

\item{which}{Integer, 1 or 2. If 1, plot W1 (input layer), if 2, plot W2 (output layer)}
}
\description{
Plot word2vec model
}
\examples{
w <- w2v(bog, hidden_dim = 2, epochs = 10, learning_rate = 0.01, verbose = 1)
plot.w2v(w, which = 1)
}
