% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w2v_nn.R
\name{w2v}
\alias{w2v}
\title{Word2vec Neural Network}
\usage{
w2v(
  corpus,
  win = 1,
  hidden_dim = 3,
  epochs = 100,
  learning_rate = 0.01,
  verbose = 10,
  weights = NULL
)
}
\arguments{
\item{corpus}{Vector of words}

\item{win}{Integer, size of the context window}

\item{hidden_dim}{Dimension of the hidden layer}

\item{epochs}{Number of training epochs}

\item{learning_rate}{Learning rate}

\item{verbose}{Integer, verbosity level (0 = no output, >0 = output every `verbose` epochs)}

\item{weights}{FIXME}
}
\description{
Word2vec Neural Network
}
\examples{
w <- w2v(bog, hidden_dim = 2, epochs = 10, learning_rate = 0.01, verbose = 1)

}
