% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w2v_nn.R
\name{w2v_build_data}
\alias{w2v_build_data}
\title{Build word2vec data from a corpus represented as a vector of words}
\usage{
w2v_build_data(corpus, win = 1)
}
\arguments{
\item{corpus}{vector of words}

\item{win}{integer, size of the context window}
}
\value{
A list containing the following elements:
- `X`: Sparse matrix of one-hot encoded input vectors
- `Y`: Sparse matrix of one-hot encoded output vectors
- `vocab`: Vector of unique words in the corpus
- `pairs`: Data frame of word pairs in context
}
\description{
Build word2vec data from a corpus represented as a vector of words
}
\examples{
w2v_data <- w2v_build_data(bog, win = 1)
}
